long change

+++ 1A.cpp
@@ -2,7 +2,7 @@
 using namespace std;
 
 int main(){
-    int n, m, a;
+    long n, m, a;
     cin >> n >> m >> a;
     int ans = ceil((float)n/a) * ceil((float)m/a);
     cout << ans << endl;
